refactor: move duplicated methods of ClassUnit and InterfaceUnit into Datatype
