+++ core/framework/common/src/main/java/edu/kit/kastel/mcse/ardoco/core/api/models/code/ClassUnit.java
@@ -2,12 +2,8 @@
 package edu.kit.kastel.mcse.ardoco.core.api.models.code;
 
 import java.io.Serial;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.SortedSet;
 
-import com.fasterxml.jackson.annotation.JsonGetter;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
 /**
@@ -19,16 +15,12 @@ public final class ClassUnit extends Datatype {
     @Serial
     private static final long serialVersionUID = 354013115794534271L;
 
-    @JsonProperty
-    private final List<String> content;
-
     /**
      * Default constructor for Jackson.
      */
     @SuppressWarnings("unused")
     private ClassUnit() {
         // Jackson
-        this.content = new ArrayList<>();
     }
 
     /**
@@ -39,11 +31,7 @@ public final class ClassUnit extends Datatype {
      * @param content            the content of the class unit
      */
     public ClassUnit(CodeItemRepository codeItemRepository, String name, SortedSet<? extends CodeItem> content) {
-        super(codeItemRepository, name);
-        this.content = new ArrayList<>();
-        for (var codeItem : content) {
-            this.content.add(codeItem.getId());
-        }
+        super(codeItemRepository, name, content);
     }
 
     /**
@@ -56,46 +44,7 @@ public final class ClassUnit extends Datatype {
      * @param endLine            the 1-indexed end line in the source file, or -1 if unknown
      */
     public ClassUnit(CodeItemRepository codeItemRepository, String name, SortedSet<? extends CodeItem> content, int startLine, int endLine) {
-        super(codeItemRepository, name, startLine, endLine);
-        this.content = new ArrayList<>();
-        for (var codeItem : content) {
-            this.content.add(codeItem.getId());
-        }
-    }
-
-    /**
-     * Returns the content IDs of this class unit.
-     *
-     * @return list of content IDs
-     */
-    @JsonGetter("content")
-    public List<String> getContentIds() {
-        return new ArrayList<>(this.content);
-    }
-
-    /**
-     * Returns the content of this class unit as a list of code items.
-     *
-     * @return list of code items
-     */
-    @Override
-    public List<CodeItem> getContent() {
-        return this.codeItemRepository.getCodeItemsByIds(this.content);
-    }
-
-    /**
-     * Returns all datatypes contained in this class unit, including itself and all nested datatypes.
-     *
-     * @return list of all datatypes
-     */
-    @Override
-    public List<Datatype> getAllDataTypes() {
-        List<Datatype> result = new ArrayList<>();
-        result.add(this);
-        for (CodeItem codeItem : this.getContent()) {
-            result.addAll(codeItem.getAllDataTypes());
-        }
-        return result;
+        super(codeItemRepository, name, content, startLine, endLine);
     }
 
     @Override
@@ -103,15 +52,6 @@ public final class ClassUnit extends Datatype {
         if (this == o) {
             return true;
         }
-        if (!(o instanceof ClassUnit classUnit) || !super.equals(o)) {
-            return false;
-        }
-        return this.content.equals(classUnit.content);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = super.hashCode();
-        return 31 * result + this.content.hashCode();
+        return o instanceof ClassUnit && super.equals(o);
     }
 }
